+++ app/src/main/kotlin/ac/mdiq/podcini/config/settings/ComboTransporter.kt
@@ -30,94 +30,94 @@ import java.io.InputStream
 import java.io.OutputStream
 import java.nio.channels.FileChannel
 
-class PreferencesTransporter(val prefsDirName: String) {
-    val TAG = "PreferencesTransporter"
-
-    @Throws(IOException::class)
-    fun exportToDocument(uri: Uri) {
-        try {
-            val chosenDir = DocumentFile.fromTreeUri(getAppContext(), uri) ?: throw IOException("Destination directory is not valid")
-            val exportSubDir = chosenDir.createDirectory(prefsDirName) ?: throw IOException("Error creating subdirectory $prefsDirName")
-            val sharedPreferencesDir = getAppContext().filesDir.parentFile?.listFiles { file -> file.name.startsWith("shared_prefs") }?.firstOrNull()
-            if (sharedPreferencesDir != null) {
-                sharedPreferencesDir.listFiles()!!.forEach { file ->
-                    val destFile = exportSubDir.createFile("text/xml", file.name)
-                    if (destFile != null) copyFile(file, destFile)
-                }
-            } else Loge("exportToDocument", "shared_prefs directory not found")
-        } catch (e: IOException) {
-            Logs(TAG, e)
-            throw e
-        } finally { }
-    }
-    private fun copyFile(sourceFile: File, destFile: DocumentFile) {
-        try {
-            val inputStream = FileInputStream(sourceFile)
-            val outputStream = getAppContext().contentResolver.openOutputStream(destFile.uri)
-            if (outputStream != null) copyStream(inputStream, outputStream)
-            inputStream.close()
-            outputStream?.close()
-        } catch (e: IOException) {
-            Logs("copyFile", e, "Error copying file:")
-            throw e
-        }
-    }
-    private fun copyFile(sourceFile: DocumentFile, destFile: File) {
-        try {
-            val inputStream = getAppContext().contentResolver.openInputStream(sourceFile.uri)
-            val outputStream = FileOutputStream(destFile)
-            if (inputStream != null) copyStream(inputStream, outputStream)
-            inputStream?.close()
-            outputStream.close()
-        } catch (e: IOException) {
-            Loge("Error", "Error copying file: $e")
-            throw e
-        }
-    }
-    private fun copyStream(inputStream: InputStream, outputStream: OutputStream) {
-        val buffer = ByteArray(1024)
-        var bytesRead: Int
-        while (inputStream.read(buffer).also { bytesRead = it } != -1) outputStream.write(buffer, 0, bytesRead)
-    }
-    @Throws(IOException::class)
-    fun importBackup(uri: Uri) {
-        try {
-            val exportedDir = DocumentFile.fromTreeUri(getAppContext(), uri) ?: throw IOException("Backup directory is not valid")
-            val sharedPreferencesDir = getAppContext().applicationContext.filesDir.parentFile?.listFiles { file -> file.name.startsWith("shared_prefs") }?.firstOrNull()
-            if (sharedPreferencesDir != null) sharedPreferencesDir.listFiles()?.forEach { file -> file.delete() }
-            else Loge("Error", "shared_prefs directory not found")
-            val files = exportedDir.listFiles()
-            var hasPodciniRPrefs = false
-            for (file in files) {
-                if (file?.isFile == true && file.name?.endsWith(".xml") == true && file.name!!.contains("podcini")) {
-                    hasPodciniRPrefs = true
-                    break
-                }
-            }
-            for (file in files) {
-                if (file?.isFile == true && file.name?.endsWith(".xml") == true) {
-                    var destName = file.name!!
-                    if (destName.contains("PlayerWidgetPrefs")) continue
-//                  for importing from Podcini version 5 and below
-                    if (!hasPodciniRPrefs) {
-                        when {
-                            destName.contains("EpisodeItemListRecyclerView") -> destName = destName.replace("EpisodeItemListRecyclerView", "EpisodesRecyclerView")
-                        }
-                    }
-                    when {
-                        BuildConfig.DEBUG && !destName.contains(".debug") -> destName = destName.replace("podcini", "podcini.debug")
-                        !BuildConfig.DEBUG && destName.contains(".debug") -> destName = destName.replace(".debug", "")
-                    }
-                    val destFile = File(sharedPreferencesDir, destName)
-                    copyFile(file, destFile)
-                }
-            }
-        } catch (e: IOException) {
-            Logs(TAG, e)
-            throw e
-        } finally { }
-    }
-}
+//class PreferencesTransporter(val prefsDirName: String) {
+//    val TAG = "PreferencesTransporter"
+//
+//    @Throws(IOException::class)
+//    fun exportToDocument(uri: Uri) {
+//        try {
+//            val chosenDir = DocumentFile.fromTreeUri(getAppContext(), uri) ?: throw IOException("Destination directory is not valid")
+//            val exportSubDir = chosenDir.createDirectory(prefsDirName) ?: throw IOException("Error creating subdirectory $prefsDirName")
+//            val sharedPreferencesDir = getAppContext().filesDir.parentFile?.listFiles { file -> file.name.startsWith("shared_prefs") }?.firstOrNull()
+//            if (sharedPreferencesDir != null) {
+//                sharedPreferencesDir.listFiles()!!.forEach { file ->
+//                    val destFile = exportSubDir.createFile("text/xml", file.name)
+//                    if (destFile != null) copyFile(file, destFile)
+//                }
+//            } else Loge("exportToDocument", "shared_prefs directory not found")
+//        } catch (e: IOException) {
+//            Logs(TAG, e)
+//            throw e
+//        } finally { }
+//    }
+//    private fun copyFile(sourceFile: File, destFile: DocumentFile) {
+//        try {
+//            val inputStream = FileInputStream(sourceFile)
+//            val outputStream = getAppContext().contentResolver.openOutputStream(destFile.uri)
+//            if (outputStream != null) copyStream(inputStream, outputStream)
+//            inputStream.close()
+//            outputStream?.close()
+//        } catch (e: IOException) {
+//            Logs("copyFile", e, "Error copying file:")
+//            throw e
+//        }
+//    }
+//    private fun copyFile(sourceFile: DocumentFile, destFile: File) {
+//        try {
+//            val inputStream = getAppContext().contentResolver.openInputStream(sourceFile.uri)
+//            val outputStream = FileOutputStream(destFile)
+//            if (inputStream != null) copyStream(inputStream, outputStream)
+//            inputStream?.close()
+//            outputStream.close()
+//        } catch (e: IOException) {
+//            Loge("Error", "Error copying file: $e")
+//            throw e
+//        }
+//    }
+//    private fun copyStream(inputStream: InputStream, outputStream: OutputStream) {
+//        val buffer = ByteArray(1024)
+//        var bytesRead: Int
+//        while (inputStream.read(buffer).also { bytesRead = it } != -1) outputStream.write(buffer, 0, bytesRead)
+//    }
+//    @Throws(IOException::class)
+//    fun importBackup(uri: Uri) {
+//        try {
+//            val exportedDir = DocumentFile.fromTreeUri(getAppContext(), uri) ?: throw IOException("Backup directory is not valid")
+//            val sharedPreferencesDir = getAppContext().applicationContext.filesDir.parentFile?.listFiles { file -> file.name.startsWith("shared_prefs") }?.firstOrNull()
+//            if (sharedPreferencesDir != null) sharedPreferencesDir.listFiles()?.forEach { file -> file.delete() }
+//            else Loge("Error", "shared_prefs directory not found")
+//            val files = exportedDir.listFiles()
+//            var hasPodciniRPrefs = false
+//            for (file in files) {
+//                if (file?.isFile == true && file.name?.endsWith(".xml") == true && file.name!!.contains("podcini")) {
+//                    hasPodciniRPrefs = true
+//                    break
+//                }
+//            }
+//            for (file in files) {
+//                if (file?.isFile == true && file.name?.endsWith(".xml") == true) {
+//                    var destName = file.name!!
+//                    if (destName.contains("PlayerWidgetPrefs")) continue
+////                  for importing from Podcini version 5 and below
+//                    if (!hasPodciniRPrefs) {
+//                        when {
+//                            destName.contains("EpisodeItemListRecyclerView") -> destName = destName.replace("EpisodeItemListRecyclerView", "EpisodesRecyclerView")
+//                        }
+//                    }
+//                    when {
+//                        BuildConfig.DEBUG && !destName.contains(".debug") -> destName = destName.replace("podcini", "podcini.debug")
+//                        !BuildConfig.DEBUG && destName.contains(".debug") -> destName = destName.replace(".debug", "")
+//                    }
+//                    val destFile = File(sharedPreferencesDir, destName)
+//                    copyFile(file, destFile)
+//                }
+//            }
+//        } catch (e: IOException) {
+//            Logs(TAG, e)
+//            throw e
+//        } finally { }
+//    }
+//}
 
 class MediaFilesTransporter(val mediaFilesDirName: String) {
     val TAG = "MediaFilesTransporter"